New helper: transformer
This new helper is basically a different and (IMHO) better approach to what groupColumnsToObjects tried to offer.

The API is different, though, so instead of breaking compatibility with the old one, I instead opted to keep them both.

Going forward `groupColumnsToObjects` will be kept the same (improvements are not planned) and the new transformer API will also be provided (with updates and improvements) separately through the `transformer` helper.

As always, 100% code coverage is guaranteed.

And documentation will soon be available.